+++ code/sound/movingpositional/movingsoundmanager.cpp
@@ -105,7 +105,14 @@ void MovingSoundManager::HandleEvent( EventEnum id, void* pEventData )
     AnimSoundDSGData* soundDSGData;
     Vehicle* theCar;
 
-    switch( id )
+    // Switched as int rather than EventEnum: EVENT_LOCATOR + LocatorEvent::X
+    // below is a deliberately valid but unnamed EventEnum value (locator
+    // sub-events are reserved a range starting at EVENT_LOCATOR, see
+    // eventenum.h's EVENT_PLACEHOLDER), not a bug - but Clang's -Wswitch
+    // flags any case value that doesn't match a named enumerator of the
+    // switch's type, which a plain int switch doesn't check (same fix as
+    // the other HandleEvent overrides in this codebase).
+    switch( static_cast<int>( id ) )
     {
         case EVENT_CHASE_VEHICLE_SPAWNED:
             //
