+++ code/worldsim/ped/pedestrianmanager.cpp
@@ -1208,9 +1208,16 @@ void PedestrianManager::HandleEvent( EventEnum id, void* pEventData )
     }
     *** DEBUG ***/
 
-    // Handle the events appropriately: 
+    // Handle the events appropriately:
     // - determine which model group we want to switch to and call SwitchToGroup on it
-    switch ( id )
+    // Switched as int rather than EventEnum: EVENT_LOCATOR + LocatorEvent::X
+    // below is a deliberately valid but unnamed EventEnum value (locator
+    // sub-events are reserved a range starting at EVENT_LOCATOR, see
+    // eventenum.h's EVENT_PLACEHOLDER), not a bug - but Clang's -Wswitch
+    // flags any case value that doesn't match a named enumerator of the
+    // switch's type, which a plain int switch doesn't check (same fix as
+    // the other HandleEvent overrides in this codebase).
+    switch ( static_cast<int>( id ) )
     {
     case EVENT_LOCATOR + LocatorEvent::LOAD_PED_MODEL_GROUP:
         {
